extract 2D XY to Flat XY conversion to own class
